+++ src/lib.rs
@@ -151,7 +151,7 @@ pub mod metrics;
 
 // Core handle + configuration.
 pub use api::builder::TreeBuilder;
-pub use api::config::{Storage, TreeConfig};
+pub use api::config::{AccessMode, Storage, TreeConfig};
 pub use api::db::{DBAtomicBatch, DBView, DB};
 pub use api::errors::{Error, Result};
 pub use api::key::{KeyPathBuf, KeyPathError, KeyPrefixBuf};